frontend: Move ReleaseNotes parts of App into ReleaseNotes.tsx
